+++ src/Products/PloneMeeting/content/meeting.py
@@ -160,13 +160,11 @@ class ICommitteesRowSchema(Interface):
             vocabulary="Products.PloneMeeting.vocabularies.selectable_committee_attendees_vocabulary"),
         required=False)
 
-    form.widget('committee_observations', PMTextAreaFieldWidget)
+    # called "committee_observations" because "observations" already exists on meeting class
     committee_observations = RichText(
         title=_(u"title_committees_committee_observations"),
-        default_mime_type='text/plain',
-        allowed_mime_types=("text/plain", ),
-        output_mime_type='text/x-html-safe',
-        required=False)
+        required=False,
+        allowed_mime_types=(u"text/html", ))
 
 
 class IMeeting(IDXMeetingContent):
@@ -1117,6 +1115,13 @@ class Meeting(Container):
             # keys are values, values are keys
             res = {v: k for k, v in res.items()}
         return res
+
+    def get_committee_observations(self, row_id, for_display=True, mark_empty_tags=False, raw=True):
+        """Return "committee_observations" for given p_row_id committee."""
+        value = self.get_committee(row_id)["committee_observations"]
+        if not value:
+            return value
+        return raw and value.raw or value.output
 
     def get_committee_items(self, row_id, supplement=-1, ordered=True, **kwargs):
         """Return every items of a given committee p_row_id.